feat: update mods, rekres, and combine everything under `talos-backup`
This PR ensures that what was previously the etcd backup service that Phillip wrote, is now the only feature in the talos-backup service. We'll expect that this code will run inside of a Talos cluster and make use of the talos api internally. Once this is merged, I'll run through testing to ensure it's working properly and, if so, cut an alpha release.
