+++ cmd/talos-backup/main.go
@@ -1,4 +1,8 @@
-// Package main provides the command line interface for the talos-backup tool.
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/.
+
+// Package main provides the command line interface for talos-backup.
 package main
 
 import (
@@ -6,98 +10,36 @@ import (
 	"fmt"
 	"log"
 	"os"
-	"time"
 
-	"github.com/adhocore/gronx"
+	talosclient "github.com/siderolabs/talos/pkg/machinery/client"
+	talosconfig "github.com/siderolabs/talos/pkg/machinery/client/config"
 
+	"github.com/siderolabs/talos-backup/cmd/talos-backup/service"
 	"github.com/siderolabs/talos-backup/pkg/config"
-	"github.com/siderolabs/talos-backup/pkg/s3"
-	"github.com/siderolabs/talos-backup/pkg/talos"
 )
 
-func main() {
+func run() error {
 	ctx := context.Background()
 
-	snapList, err := config.ParseConfig(ctx, "./config.yaml")
+	serviceConfig := config.GetServiceConfig()
+
+	talosConfig, err := talosconfig.Open("")
 	if err != nil {
-		log.Fatal(err)
+		return fmt.Errorf("failed to get talosconfig: %w", err)
 	}
 
-	ticker := time.NewTicker(60 * time.Second)
-
-	for range ticker.C {
-		errs := make(chan error, len(snapList.Snapshots))
-
-		for _, snapItem := range snapList.Snapshots {
-			go processSnap(ctx, snapItem, errs)
-		}
-
-		for range snapList.Snapshots {
-			err := <-errs
-			if err != nil {
-				log.Printf("%v", err)
-			}
-		}
-
-		close(errs)
+	talosClient, err := talosclient.New(ctx, talosclient.WithConfig(talosConfig))
+	if err != nil {
+		return fmt.Errorf("failed to create talos client: %w", err)
 	}
+
+	return service.BackupEncryptedSnapshot(ctx, serviceConfig, talosConfig, talosClient)
 }
 
-func processSnap(ctx context.Context, snapItem config.Snapshot, errChan chan error) {
-	gron := gronx.New()
+func main() {
+	if err := run(); err != nil {
+		log.Println(err)
 
-	isDue, err := gron.IsDue(snapItem.Schedule)
-	if err != nil {
-		errChan <- fmt.Errorf("failed validating schedule for cluster %q: %w", snapItem.ClusterName, err)
-
-		return
+		os.Exit(-1)
 	}
-
-	if isDue {
-		log.Printf("cluster %q due for snapshotting", snapItem.ClusterName)
-
-		err = snapIt(ctx, snapItem)
-		if err != nil {
-			errChan <- err
-
-			return
-		}
-	}
-	errChan <- nil
-}
-
-func snapIt(ctx context.Context, snapItem config.Snapshot) error {
-	tc, err := talos.CreateClient(ctx, snapItem.TalosInfo.Secret)
-	if err != nil {
-		return fmt.Errorf("failed creating talos client for cluster %q: %w", snapItem.ClusterName, err)
-	}
-
-	log.Printf("taking etcd snapshot for cluster %q", snapItem.ClusterName)
-
-	snapPath, err := talos.TakeEtcdSnapshot(ctx, tc, snapItem.ClusterName)
-	if err != nil {
-		return fmt.Errorf("failed taking snapshot for cluster %q: %w", snapItem.ClusterName, err)
-	}
-
-	s3c, err := s3.CreateClient(ctx, snapItem.S3Info)
-	if err != nil {
-		return fmt.Errorf("failed creating s3 client for cluster %q: %w", snapItem.ClusterName, err)
-	}
-
-	log.Printf("pushing snapshot %q for cluster %q to s3 bucket %q", snapPath, snapItem.ClusterName, snapItem.S3Info.Bucket)
-
-	err = s3.PushSnapshot(ctx, snapItem.S3Info, s3c, snapItem.ClusterName, snapPath)
-	if err != nil {
-		return fmt.Errorf("failed pushing snapshot to s3 for cluster %q: %w", snapItem.ClusterName, err)
-	}
-
-	// Only after pushing will we delete the local copy. this might cause heartburn if we can never push for some reason :shrug:
-	log.Printf("cleaning up pushed snapshot %q for cluster %q", snapPath, snapItem.ClusterName)
-
-	err = os.Remove(snapPath)
-	if err != nil {
-		return fmt.Errorf("failed cleaning up pushed snapshot to s3 for cluster %q: %w", snapItem.ClusterName, err)
-	}
-
-	return nil
 }
